code_reviewer/requirements.txt
google-genai
python-dotenv
colorama
requests